requirements.txt
streamlit==1.44.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
python-dotenv==1.1.0
faiss-cpu==1.10.0
PyMuPDF==1.23.26
pandas==2.2.3
requests==2.32.3
PyPDF2